:wrench: Strip unused sections from assembly
The [Rust core library](https://doc.rust-lang.org/core/) includes some
things which we don't. We can't even compile because of it.

The `--gc-sections` flag passed to `ld` strips extra sections if we
don't *seem* to use them. Because we don't *seem* to use the
`.multiboot_header` section, we have to instruct `ld` to keep it anyway.